+++ src/Plugin/Plan/PlanType/PracticeImplementation.php
@@ -49,6 +49,12 @@ class PracticeImplementation extends FarmPlanType {
         'required' => TRUE,
       ],
 
+      // Geometry.
+      'geometry' => [
+        'type' => 'geofield',
+        'label' => $this->t('Geometry'),
+      ],
+
       // Conservation practice.
       'rcd_practice' => [
         'type' => 'list_string',